Separated MovieGrid functions to their own files
